fix some misspell,and size-t for grade transform on int

+++ ex00/include/Bureaucrat.hpp
@@ -6,7 +6,7 @@
 /*   By: tsaby <tsaby@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/23 11:03:13 by tsaby             #+#    #+#             */
-/*   Updated: 2026/01/23 12:03:11 by tsaby            ###   ########.fr       */
+/*   Updated: 2026/06/02 18:54:05 by tsaby            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ class Bureaucrat
 {
 	private:
 		const std::string _name;
-		size_t _rank;
+		int _rank;
 	public:
 		Bureaucrat();
-		Bureaucrat(const std::string& name, size_t rank);
+		Bureaucrat(const std::string& name, int rank);
 		Bureaucrat(const Bureaucrat &copy);
 		Bureaucrat& operator=(const Bureaucrat& other);
 		~Bureaucrat();
@@ -38,7 +38,7 @@ class Bureaucrat
 		};
 
 		std::string getName() const;
-		size_t getGrade() const;
+		int getGrade() const;
 		void 	incrementGrade();
 		void 	decrementGrade();
 };
